BACK/requirements.txt
appdirs==1.4.4
fastapi==0.115.6
jsonpointer==2.1
setuptools==75.1.0
starlette==0.41.3
uvicorn==0.32.1
PyJWT==2.3.0   
bcrypt
mysql-connector-python==8.2.0
mariadb
requests
apscheduler
httpx
pandas
numpy
tqdm
shapely
